perf(plugins): keep bundled catalog lookups lightweight
Share the typed raw bundled inventory, short-circuit direct lookups, and keep hosted fetching and artifact download code behind their runtime owners. Preserve live row mutation, fresh list arrays, source-reference filtering, conservative raw repair targets, hosted trust/freshness and integrity behavior.

Validated 138 focused tests, full check:changed, full build, and independent isolated P2 autoreview. In an isolated built probe, 95,000 lookups took 531.07 ms instead of 1001.76 ms, and cold import loaded 58 modules instead of 255; these are catalog measurements, not full-Gateway claims.
