+++ src/infra/clawhub-artifacts.ts
@@ -2,10 +2,7 @@
 import { createHash } from "node:crypto";
 import fs from "node:fs/promises";
 import { parseStrictPositiveInteger } from "@openclaw/normalization-core/number-coercion";
-import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-} from "@openclaw/normalization-core/string-coerce";
+import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import {
   createClawHubError,
   readClawHubBytes,
@@ -13,6 +10,7 @@ import {
   resolveClawHubBaseUrl,
   type ClawHubFetch,
 } from "./clawhub-client.js";
+import { normalizeClawHubSha256Hex } from "./clawhub-integrity.js";
 import { sha256Base64, sha256Hex } from "./crypto-digest.js";
 import { createTempDownloadTarget } from "./temp-download.js";
 
@@ -90,43 +88,6 @@ async function stageClawHubArchive(params: {
     await target.cleanup().catch(() => undefined);
     throw error;
   }
-}
-
-/** Normalizes ClawHub SHA-256 metadata into Subresource Integrity format. */
-export function normalizeClawHubSha256Integrity(value: string): string | null {
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return null;
-  }
-  const prefixedBase64 = /^sha256-([A-Za-z0-9+/]+={0,1})$/.exec(trimmed);
-  if (prefixedBase64?.[1]) {
-    try {
-      const decoded = Buffer.from(prefixedBase64[1], "base64");
-      if (decoded.length === 32) {
-        return `sha256-${decoded.toString("base64")}`;
-      }
-    } catch {
-      return null;
-    }
-    return null;
-  }
-  const prefixedHex = /^sha256:([A-Fa-f0-9]{64})$/.exec(trimmed);
-  if (prefixedHex?.[1]) {
-    return `sha256-${Buffer.from(prefixedHex[1], "hex").toString("base64")}`;
-  }
-  if (/^[A-Fa-f0-9]{64}$/.test(trimmed)) {
-    return `sha256-${Buffer.from(trimmed, "hex").toString("base64")}`;
-  }
-  return null;
-}
-
-/** Normalizes ClawHub SHA-256 metadata into lowercase hex form. */
-export function normalizeClawHubSha256Hex(value: string): string | null {
-  const trimmed = value.trim();
-  if (!/^[A-Fa-f0-9]{64}$/.test(trimmed)) {
-    return null;
-  }
-  return normalizeLowercaseStringOrEmpty(trimmed);
 }
 
 export async function downloadClawHubPackageArchive(params: {
