+++ src/plugins/official-external-plugin-targets.ts
@@ -3,39 +3,7 @@ import { isRecord } from "@openclaw/normalization-core/record-coerce";
 import { normalizeOptionalLowercaseString } from "@openclaw/normalization-core/string-coerce";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import { BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES } from "./official-external-plugin-bundled-catalogs.js";
-
-type StaticProvider = {
-  id?: string;
-  aliases?: readonly string[];
-  envVars?: readonly string[];
-};
-
-type StaticWebProvider = {
-  id?: string;
-  envVars?: readonly string[];
-};
-
-type StaticConfiguredState = {
-  env?: {
-    allOf?: readonly string[];
-    anyOf?: readonly string[];
-  };
-};
-
-type StaticManifest = {
-  channel?: {
-    id?: string;
-    envVars?: readonly string[];
-    configuredState?: StaticConfiguredState;
-  };
-  contracts?: Record<string, readonly string[]>;
-  providers?: readonly StaticProvider[];
-  webSearchProviders?: readonly StaticWebProvider[];
-};
-
-type StaticEntry = { openclaw?: StaticManifest };
-
-const STATIC_ENTRIES = BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES as readonly StaticEntry[];
+import type { OfficialExternalPluginCatalogManifest } from "./official-external-plugin-catalog.types.js";
 
 function normalizeIds(values: Iterable<string>): Set<string> {
   return new Set(
@@ -51,7 +19,7 @@ function envHasAny(env: NodeJS.ProcessEnv, names: readonly string[] | undefined)
 
 function envHasChannelCandidate(
   env: NodeJS.ProcessEnv,
-  channel: StaticManifest["channel"],
+  channel: OfficialExternalPluginCatalogManifest["channel"],
 ): boolean {
   const allOf = channel?.configuredState?.env?.allOf ?? [];
   const anyOf = channel?.configuredState?.env?.anyOf ?? [];
@@ -63,7 +31,7 @@ export function hasOfficialExternalProviderTarget(params: {
   env: NodeJS.ProcessEnv;
 }): boolean {
   const providerIds = normalizeIds(params.providerIds);
-  return STATIC_ENTRIES.some((entry) =>
+  return BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES.some((entry) =>
     entry.openclaw?.providers?.some(
       (provider) =>
         envHasAny(params.env, provider.envVars) ||
@@ -76,14 +44,14 @@ export function hasOfficialExternalProviderTarget(params: {
 }
 
 export function hasOfficialExternalContractTarget(params: {
-  contract: string;
+  contract: keyof NonNullable<OfficialExternalPluginCatalogManifest["contracts"]>;
   providerIds: Iterable<string>;
 }): boolean {
   const providerIds = normalizeIds(params.providerIds);
   if (providerIds.size === 0) {
     return false;
   }
-  return STATIC_ENTRIES.some((entry) =>
+  return BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES.some((entry) =>
     entry.openclaw?.contracts?.[params.contract]?.some((providerId) => {
       const normalized = normalizeOptionalLowercaseString(providerId);
       return normalized ? providerIds.has(normalized) : false;
@@ -92,10 +60,10 @@ export function hasOfficialExternalContractTarget(params: {
 }
 
 export function hasOfficialExternalWebContractEnvTarget(params: {
-  contract: string;
+  contract: keyof NonNullable<OfficialExternalPluginCatalogManifest["contracts"]>;
   env: NodeJS.ProcessEnv;
 }): boolean {
-  return STATIC_ENTRIES.some((entry) => {
+  return BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES.some((entry) => {
     const manifest = entry.openclaw;
     const contractIds = normalizeIds(manifest?.contracts?.[params.contract] ?? []);
     return manifest?.webSearchProviders?.some((provider) => {
@@ -112,7 +80,7 @@ export function hasOfficialExternalChannelTarget(params: {
   env: NodeJS.ProcessEnv;
 }): boolean {
   const channels = isRecord(params.config.channels) ? params.config.channels : undefined;
-  return STATIC_ENTRIES.some((entry) => {
+  return BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES.some((entry) => {
     const channel = entry.openclaw?.channel;
     const channelId = normalizeOptionalLowercaseString(channel?.id);
     if (!channelId) {
@@ -131,7 +99,7 @@ export function hasOfficialExternalWebSearchTarget(params: {
   env: NodeJS.ProcessEnv;
 }): boolean {
   const configuredId = normalizeOptionalLowercaseString(params.providerId);
-  return STATIC_ENTRIES.some((entry) =>
+  return BUNDLED_OFFICIAL_EXTERNAL_PLUGIN_CATALOG_ENTRIES.some((entry) =>
     entry.openclaw?.webSearchProviders?.some((provider) => {
       const providerId = normalizeOptionalLowercaseString(provider.id);
       return (
